refactor redis into its own class too
